feat(workflow): canonicalize workflow route and SMS approval
